agent_services/python/requirements.txt
grpcio==1.0.4
grpcio-tools==1.0.4
inject==3.3.1
